requirements.txt
asyncio==3.4.3
av==14.4.0
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.58.4
humanfriendly==10.0
kiwisolver==1.4.8
lap==0.5.12
loguru==0.7.3
matplotlib==3.10.3
mpmath==1.3.0
numpy==2.3.1
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
packaging==25.0
paho-mqtt==2.1.0
pillow==11.2.1
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rtmlib==0.0.13
scipy==1.16.0
six==1.17.0
sympy==1.14.0
tqdm==4.67.1
